+++ src/lib_smart_rollup_node/daemon_event.ml
@@ -163,6 +163,15 @@ module Simple = struct
       ("delay", Time.System.Span.encoding)
       ~pp1:Ptime.Span.pp
 
+  let main_loop_retry =
+    declare_1
+      ~section
+      ~name:"smart_rollup_node_daemon_main_loop_retry"
+      ~msg:"Restarting main rollup node loop in {delay}."
+      ~level:Warning
+      ("delay", Time.System.Span.encoding)
+      ~pp1:Ptime.Span.pp
+
   let exit_bailout_mode =
     declare_0
       ~section
@@ -219,5 +228,8 @@ let degraded_mode () = Simple.(emit degraded_mode) ()
 
 let refutation_loop_retry d =
   Simple.(emit refutation_loop_retry) (Time.System.Span.of_seconds_exn d)
+
+let main_loop_retry d =
+  Simple.(emit main_loop_retry) (Time.System.Span.of_seconds_exn d)
 
 let exit_bailout_mode () = Simple.(emit exit_bailout_mode) ()